+++ framework/directives/pullHook.js
@@ -114,6 +114,14 @@
 
 /**
  * @ngdoc attribute
+ * @name fixed-content
+ * @description
+ *   [en]If this attribute is set the content of the page will not move when pulling.[/en]
+ *   [ja]この属性がある時、プルフックが引き出されている時にもコンテンツは動きません。[/ja]
+ */
+
+/**
+ * @ngdoc attribute
  * @name ons-changestate
  * @type {Expression}
  * @description