+++ CHANGELOG.md
@@ -15,6 +15,7 @@ v1.3.1-dev
  * ons-list-item: Added `tight` modifier.
  * ons-if-platform: Supports multiple space separated values.
  * ons-pull-hook: Added `getPullDistance`, `getHeight`, `getThresholdHeight` methods.
+ * ons-pull-hook: Added `fixed-content` attribute.
 
 v1.3.0
 ----